# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VirtualPC Admin API \u2014 Auth/Audit/Dashboard/Security",
    "version": "1.0.0",
    "description": "Role-based admin API. Bearer auth via sessionId from POST /api/auth/login."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "sessionId returned by /api/auth/login"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string",
            "description": "Human-readable error message."
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "ceo"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "ceo123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "AuthTokenResponse": {
        "type": "object",
        "description": "Successful authentication. The `token` is the sessionId to send as 'Authorization: Bearer <token>'. Sessions expire after 8 hours.",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ],
            "example": true
          },
          "token": {
            "type": "string",
            "description": "Session ID / Bearer token.",
            "example": "session_1717400000000_a1b2c3d4e"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserSummary"
          }
        },
        "required": [
          "success",
          "token",
          "user"
        ]
      },
      "TwoFactorRequiredResponse": {
        "type": "object",
        "description": "Returned by POST /api/auth/login with HTTP 200 when the account has 2FA enabled. No session is issued; exchange the challengeId via POST /api/auth/2fa/verify. Challenge expires after 5 minutes and is single-use.",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ],
            "example": false
          },
          "requires2fa": {
            "type": "boolean",
            "enum": [
              true
            ],
            "example": true
          },
          "challengeId": {
            "type": "string",
            "example": "2fa_1717400000000_abcdef0123456789abcdef01"
          }
        },
        "required": [
          "success",
          "requires2fa",
          "challengeId"
        ]
      },
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Login with username and password",
        "description": "Authenticates a user and issues a Bearer session token (valid 8h). Rate limited to 10 attempts per 15 minutes per (IP, username) pair, on top of a per-username brute-force lockout (5 failures -> 15 min cooldown). If the account has 2FA enabled the password check still passes but no session is issued: the response is HTTP 200 with { success:false, requires2fa:true, challengeId } and the client must call POST /api/auth/2fa/verify with the challengeId and a TOTP code.",
        "tags": [
          "Auth"
        ],
        "security": [],
        "parameters": [
          {
            "name": "X-Device-Id",
            "in": "header",
            "required": false,
            "description": "Client device identifier, recorded against the session / 2FA challenge. Defaults to 'unknown-device'.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Location",
            "in": "header",
            "required": false,
            "description": "Client location hint, recorded against the session / 2FA challenge. Defaults to 'unknown-location'.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Either a successful login (session token issued) OR a 2FA challenge requiring a second step. The HTTP status is 200 in both cases; distinguish via the `success` and `requires2fa` fields.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/AuthTokenResponse"
                    },
                    {
                      "$ref": "#/components/schemas/TwoFactorRequiredResponse"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials, inactive account, or brute-force lockout.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (10 login attempts per 15 minutes per IP+username).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/2fa/verify": {
      "post": {
        "summary": "Complete a 2FA-required login",
        "description": "Exchanges the single-use challengeId returned by POST /api/auth/login (when requires2fa is true) together with a 6-digit TOTP code for a real session token. The challenge is consumed whether or not the code matches and expires 5 minutes after issuance. Shares the login rate limiter (10 attempts / 15 min per IP+username).",
        "tags": [
          "Auth",
          "2FA"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA verified; session token issued.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid/expired challenge, invalid 2FA state, or invalid TOTP code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/2fa/setup": {
      "post": {
        "summary": "Begin 2FA (TOTP) setup",
        "description": "Generates a fresh TOTP secret and otpauth:// URI for the authenticated user to render as a QR code. Must be followed by POST /api/auth/2fa/enable to actually arm 2FA. Calling setup again before enable rotates the secret. Rate limited to 30 mutations/min per IP.",
        "tags": [
          "Auth",
          "2FA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Setup initiated; returns the secret and otpauth URI.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TwoFactorSetupResponse"
                }
              }
            }
          },
          "400": {
            "description": "Setup could not be initiated (e.g. user not found).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/2fa/enable": {
      "post": {
        "summary": "Confirm and arm 2FA",
        "description": "Confirms TOTP setup by submitting a code generated from the secret returned by POST /api/auth/2fa/setup, then arms 2FA on the account. Fails if setup has not been run. Rate limited to 30 mutations/min per IP.",
        "tags": [
          "Auth",
          "2FA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorCodeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA enabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Setup not run, or invalid 2FA code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/2fa/disable": {
      "post": {
        "summary": "Disable 2FA",
        "description": "Disarms 2FA for the authenticated user. Requires BOTH the current password and a valid TOTP code, so a stolen session alone cannot disable 2FA. Rate limited to 30 mutations/min per IP.",
        "tags": [
          "Auth",
          "2FA"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TwoFactorDisableRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Password incorrect or invalid 2FA code.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout / invalidate the current session",
        "description": "Invalidates the session associated with the supplied Bearer token. Idempotent.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Session invalidated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "summary": "Get current user profile",
        "description": "Returns the authenticated user's profile plus their role-based permission matrix.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile and permissions for the current user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Authenticated user record not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/change-password": {
      "post": {
        "summary": "Change the current user's password",
        "description": "Changes the authenticated user's password after verifying the old password. Rate limited to 30 mutations/min per IP.",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Current password incorrect or user not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or expired Bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/sessions": {
      "get": {
        "summary": "Get session statistics (CEO & CTO only)",
        "description": "Returns aggregate session and user statistics including active session counts and per-role breakdowns. Restricted to CEO and CTO roles via requireRole('ceo','cto').",
        "tags": [
          "Auth",
          "Sessions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Session statistics.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing Bearer token (requireRole returns 401 when req.user is unset).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but not a CEO or CTO.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}